+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- Added user preferences [PR#144](https://github.com/silvioubaldino/personal-finance/pull/144)
 - Added internal transfer [PR#143](https://github.com/silvioubaldino/personal-finance/pull/143)
 - Fixed update wallet in revert invoice payment [PR#142](https://github.com/silvioubaldino/personal-finance/pull/142)