+++ CHANGELOG.md
@@ -8,6 +8,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+-
+
+## Release - v1.7.0 - 30-12-2025
+
 - Improved user preferences flow [PR#145](https://github.com/silvioubaldino/personal-finance/pull/145)
 - Added user preferences [PR#144](https://github.com/silvioubaldino/personal-finance/pull/144)
 - Added internal transfer [PR#143](https://github.com/silvioubaldino/personal-finance/pull/143)